✨ Update Popover with content Event Card
